+++ app/States/Posts/PostUnpublished.php
@@ -2,7 +2,9 @@
 
 namespace App\States\Posts;
 
+use App\Features\PublishPostToX;
 use App\Jobs\PostToXJob;
+use Laravel\Pennant\Feature;
 
 class PostUnpublished extends PostState
 {
@@ -17,7 +19,10 @@ class PostUnpublished extends PostState
 		if (!$this->post->published_at) {
 			$this->post->published_at = now();
 
-			PostToXJob::dispatch($this->post->getPostable());
+			PostToXJob::dispatchIf(
+				Feature::active(PublishPostToX::class),
+				$this->post->getPostable(),
+			);
 		}
 
 		return $this->post->save() || parent::publish();